+++ include/rak/compiler.h
@@ -23,6 +23,6 @@ typedef struct
 void rak_compiler_init(RakCompiler *comp, RakError *err);
 void rak_compiler_deinit(RakCompiler *comp);
 void rak_compiler_compile_chunk(RakCompiler *comp, char *source, RakError *err);
-void rak_compiler_eompile_expr(RakCompiler *comp, char *source, RakError *err);
+void rak_compiler_compile_expr(RakCompiler *comp, char *source, RakError *err);
 
 #endif // RAK_COMPILER_H