+++ src/main.c
@@ -18,13 +18,14 @@
 static RakError err;
 static RakString source;
 static RakCompiler comp;
+static RakVM vm;
 
 static inline void check_error(void);
 static inline void init(void);
 static inline void deinit(void);
 static inline void repl(void);
 static inline bool read(void);
-static inline void compile(void);
+static inline void eval(void);
 
 static inline void check_error(void)
 {
@@ -40,12 +41,15 @@ static inline void init(void)
   check_error();
   rak_compiler_init(&comp, &err);
   check_error();
+  rak_vm_init(&vm, RAK_VM_VSTK_DEFAULT_SIZE, &err);
+  check_error();
 }
 
 static inline void deinit(void)
 {
   rak_string_deinit(&source);
   rak_compiler_deinit(&comp);
+  rak_vm_deinit(&vm);
 }
 
 static inline void repl(void)
@@ -55,7 +59,7 @@ static inline void repl(void)
     printf("> ");
     if (!read()) break;
     if (source.slice.len == 1) continue;
-    compile();
+    eval();
   }
 }
 
@@ -68,16 +72,25 @@ static inline bool read(void)
   return true;
 }
 
-static inline void compile(void)
+static inline void eval(void)
 {
   err = rak_ok();
-  rak_compiler_compile_chunk(&comp, rak_string_chars(&source), &err);
+  rak_compiler_compile_expr(&comp, rak_string_chars(&source), &err);
   if (!rak_is_ok(&err))
   {
     rak_error_print(&err);
     return;
   }
-  rak_dump_chunk(&comp.chunk);
+  rak_vm_run(&vm, &comp.chunk, &err);
+  if (!rak_is_ok(&err))
+  {
+    rak_error_print(&err);
+    return;
+  }
+  RakValue val = rak_vm_get(&vm, 0);
+  rak_value_print(val);
+  printf("\n");
+  rak_vm_pop(&vm);
 }
 
 int main(void)